Finished Stage Display
Stage display now displays the current stage, and if it is a foe stage, it will show the active cards in the stage, the total BP of the foe, and if the foe is boosted or not (if it is the totalBP will use the boosted value). Stage display also shows the current BP of all the "activePlayers" the players who are currently participating in the quest. NOTE: Issues in getting the png file for the background. Its just a matter of getting the file from the right place now.