Fixed when stages are displayed
Stages are now only displayed in the QuestDisplay when you are the person sponsoring the quest, and when the FoeStageStarts (i.e. you will only see the first stage once the FoeStageStarts, rather than while the other player is sponsoring the quest, while the sponsor can see all of the stages until the quest begins).