issue 103-1, Jayson
Final implementation and testing for Events.

Implemented Last event card KingsCallToArms.

First implementation and testing of Event Phase

Made some changes to Hand to support discard tracking and to multiple card areas.

Areas that call active effects now set an observer on effect that will notify it when the effect has completed. Effects may be completed but not resolved if they have a delayed trigger.